refactor(name change): changed the folder 'src/common/Macros' to 'src/common/macros' as the lowercase spelling is easier to type and looks nicer imho
