+++ src/common/macros/forward_declaration.hpp
